Move more files into the search src directory (#36575)
