+++ src/search/components/index.tsx
@@ -6,14 +6,14 @@ import { Heading } from '@primer/react'
 import { sendEvent, EventType } from 'src/events/components/events'
 import { useTranslation } from 'components/hooks/useTranslation'
 import { DEFAULT_VERSION, useVersion } from 'components/hooks/useVersion'
-import { useNumberFormatter } from 'components/hooks/useNumberFormatter'
+import { useNumberFormatter } from 'src/search/components/useNumberFormatter'
 import type { SearchResultsT } from 'src/search/components/types'
 import { SearchResults } from 'src/search/components/SearchResults'
 import { SearchError } from 'src/search/components/SearchError'
 import { NoQuery } from 'src/search/components/NoQuery'
 import { Loading } from 'src/search/components/Loading'
-import { useQuery } from 'components/hooks/useQuery'
-import { usePage } from 'components/hooks/usePage'
+import { useQuery } from 'src/search/components/useQuery'
+import { usePage } from 'src/search/components/usePage'
 import { useMainContext } from 'components/context/MainContext'
 
 export function Search() {
