+++ src/search/components/useNumberFormatter.ts
